+++ lib/gravity/NBodySolver.h
@@ -8,7 +8,6 @@
 #include "objects/containers/FlatSet.h"
 #include "thread/ThreadLocal.h"
 #include "timestepping/ISolver.h"
-#include <set>
 
 NAMESPACE_SPH_BEGIN
 
@@ -48,7 +47,7 @@ private:
     FlatSet<Size> removed;
 
     /// Holds all detected collisions.
-    std::set<CollisionRecord> collisions;
+    Array<CollisionRecord> collisions;
 
     /// Maximum distance to search for impactors, per particle.
     Array<Float> searchRadii;